+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Refactor collection text id handling.
 - Migrate constructor-based initializers to class field initializers where possible.
 - Migrate manuscripts component to zoneless-ready.
+- Migrate reading-text component to zoneless-ready.
 - The selected text title in the text-select-wrapper (facsimiles, manuscripts and variants) is enclosed in a `<div>` rather than a `<h3>` for improved semantics.
 - Deps: update `@angular/cli` to 20.2.0 and `@angular/core` to 20.2.1.
 - Deps: update `marked` to 16.2.0.